// ts-ground/ts-ground-admin-ui/lib/admin/types.ts
export interface HttpClient extends Record<string, unknown> {
  get<T>(path: string): Promise<T>;
  post<T>(path: string, data?: unknown): Promise<T>;
  put<T>(path: string, data: unknown): Promise<T>;
  patch<T>(path: string, data: unknown): Promise<T>;
  delete<T>(path: string): Promise<T>;
}